orchestrator/requirements.txt
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1